requirements.txt
beautifulsoup4==4.12.3
elasticsearch==8.12.0
gradio==4.16.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
openai==1.10.0
pydantic==2.6.0
python-decouple==3.8
tiktoken==0.5.2
unstructured==0.12.3
lark==1.1.9
